subscriptions: make changes w/concurrency
So as to process more than 1 change at a time, so as to reduce the wall-clock time required to process hundreds of repos.

Also: show a progress bar now, instead of outputting all the repo names.
When one is running these scripts on hundreds of repos, it would
probably be inconvenient to have them all printed out to the terminal.
And it wouldn’t be super-useful: I suspect the user would be more
concerned with monitoring progress.