Reorg so as to support multiple Ruby toolsets
Right now we have 3 scripts related to subscriptions; I think of that as
a single toolset for subscriptions. I’m about to add a new toolset for
working with repos, and I want to reduce duplication between these two
toolsets. So although I’m not thrilled to group toolsets by programming
system, it feels like the pragmatic thing to do here.