Make run_all_tests.sh more portable
The docker image we’re using in CircleCI uses Alpine Linux, so it
doesn’t have bash installed. Sure, we could install bash, but that’d
slow the builds down or require me to understand caching on CircleCI.

So instead I’m changing this script to be more portable across POSIX
shells.